+++ config.go
@@ -23,8 +23,6 @@ import (
 	"go.opentelemetry.io/otel/metric"
 	"go.opentelemetry.io/otel/propagation"
 	"go.opentelemetry.io/otel/trace"
-
-	internalsemconv "github.com/XSAM/otelsql/internal/semconv"
 )
 
 const (
@@ -33,8 +31,6 @@ const (
 
 var (
 	connectionStatusKey = attribute.Key("status")
-	queryStatusKey      = attribute.Key("status")
-	queryMethodKey      = attribute.Key("method")
 )
 
 // SpanNameFormatter supports formatting span names.
@@ -97,13 +93,6 @@ type config struct {
 	// Default is false
 	DisableSkipErrMeasurement bool
 
-	// SemConvStabilityOptIn controls which database semantic convention are emitted.
-	// It follows the value of environment variable `OTEL_SEMCONV_STABILITY_OPT_IN`.
-	SemConvStabilityOptIn internalsemconv.OTelSemConvStabilityOptInType
-
-	// DBQueryTextAttributes will be called to produce related attributes on `db.query.text`.
-	// It follows the value of environment variable `OTEL_SEMCONV_STABILITY_OPT_IN`.
-	DBQueryTextAttributes func(query string) []attribute.KeyValue
 }
 
 // SpanOptions holds configuration of tracing span to decide
@@ -122,7 +111,7 @@ type SpanOptions struct {
 	// DisableErrSkip, if set to true, will suppress driver.ErrSkip errors in spans.
 	DisableErrSkip bool
 
-	// DisableQuery if set to true, will suppress db.statement in spans.
+	// DisableQuery if set to true, will suppress db.query.text in spans.
 	DisableQuery bool
 
 	// RecordError, if set, will be invoked with the current error, and if the func returns true
@@ -162,11 +151,6 @@ func newConfig(options ...Option) config {
 		TracerProvider:    otel.GetTracerProvider(),
 		MeterProvider:     otel.GetMeterProvider(),
 		SpanNameFormatter: defaultSpanNameFormatter,
-		// Uses the stable behavior
-		SemConvStabilityOptIn: internalsemconv.OTelSemConvStabilityOptInStable,
-		DBQueryTextAttributes: internalsemconv.NewDBQueryTextAttributes(
-			internalsemconv.OTelSemConvStabilityOptInStable,
-		),
 	}
 	for _, opt := range options {
 		opt.Apply(&cfg)
@@ -187,12 +171,6 @@ func newConfig(options ...Option) config {
 	if cfg.Instruments, err = newInstruments(cfg.Meter); err != nil {
 		otel.Handle(err)
 	}
-
-	// Initialize SemConvStabilityOptIn from environment
-	cfg.SemConvStabilityOptIn = internalsemconv.ParseOTelSemConvStabilityOptIn()
-
-	// Initialize DBQueryTextAttributes based on SemConvStabilityOptIn
-	cfg.DBQueryTextAttributes = internalsemconv.NewDBQueryTextAttributes(cfg.SemConvStabilityOptIn)
 
 	return cfg
 }